+++ Example/Tests/Tests.swift
@@ -6,4 +6,14 @@ class Tests: XCTestCase {
     let action = UIAlertAction(title: "Hello world!", image: #imageLiteral(resourceName: "settings"), style: .cancel)
     XCTAssertNotNil(action.image)
   }
+    
+    func testCheckmarkIsCorrectlySet() {
+      let action = UIAlertAction(title: "Hello world!", image: #imageLiteral(resourceName: "settings"), isSelected: true, style: .cancel)
+      XCTAssertTrue(action.isSelected)
+    }
+    
+    func testCheckmarkIsNotSet() {
+      let action = UIAlertAction(title: "Hello world!", image: #imageLiteral(resourceName: "settings"), style: .cancel)
+      XCTAssertFalse(action.isSelected)
+    }
 }