+++ Example/ImageAlertAction/ViewController.swift
@@ -16,7 +16,7 @@ class ViewController: UIViewController {
     let settings = UIAlertAction(title: "Settings", image: #imageLiteral(resourceName: "settings"), style: .default)
     alertController.addAction(settings)
     
-    let selectedSettings = UIAlertAction(title: "Selected Settings", image: #imageLiteral(resourceName: "settings"), isSelected: true, style: .default)
+    let selectedSettings = UIAlertAction(title: "Checked Settings", image: #imageLiteral(resourceName: "settings"), isChecked: true, style: .default)
     alertController.addAction(selectedSettings)
 
     let cancel = UIAlertAction(title: "Cancel", style: .cancel)