+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## Next
+
 ## 3.2.0
 - Added support to display a checkmark with `isSelected`